add self feed, Safari support, design improvements
Add support for merging your private actor feed into your private feed
(to see your own actions.) This setting is the default, and can be
disabled with a checkbox on the back screen.

Add support for running as a file:/// URL in Safari. This involved major
updates to the CSS and js.

Visual improvements:

  * short date floats right

  * description date is a full and relative date

Add apple widget resources into project.

TOOD:

  * rewrite API calls to use the Github events API using CORS or JSONP
    so that the app works as a live web app

  * support Chrome and Firefox