update relative dates every 15s when visible
Set an interval timer to update the relative dates in the descriptions
every 15 seconds. This will present a more accurate display when there
are relative dates such as "a few seconds ago" or "1 minute ago".

Also make the feed refresh interval configurable in attributes.js in
minutes. Default is 15 minutes.