fix same date pair dedup correctly
Rewrite the feed caching and merging code so that duplicate pairs are
correctly removed.

This is done by storing each feed keyed by URL, and only adding new
entries to the feed cache. This way the same feed entries do not have to
be deduped.

deDup is still called just in case, but should no longer do anything.